Report error= tag according to spec
With this change in place, we now report `error=true` (note lower case
`t`) as described
[in the spec](https://github.com/opentracing/specification/blob/master/semantic_conventions.md#span-tags-table),
when something fails inside of a span.

Before this change, errors were reported with a capital `T`.

Then, when using the Lightstep web UI to search for `error=true`, you
won't find any Python failures.

To find the Python errors, you had to search for `error=True`, which is
not obvious.