Adding on span error callbacks
I'd like to extend the functionality of Span._on_error to allow
integrating other tools with the error handling of the spans.

I don't think monkey patching is the good way to go on this, since also
`Span._on_error` is a private method.

Another approach would be having a `exit_callbacks` list to be executed
on each `__exit__` call.

I am not sure this funcionality exists on other languages implementation
but it'd open the door to do more powerful integrtions with OpenTracing.